aiagent/requirements.txt
nats-py
psycopg2-binary
numpy
scikit-learn
openai